Minor CharType fix

+++ src/main/java/core/procedures/interop/CharType.kt
@@ -7,7 +7,7 @@ import core.procedures.Arity.Exactly
 class CharType : AFn<Any?, Char>(name = "char", isPure = true, arity = Exactly(1)) {
 
     override operator fun invoke(arg: Any?) = when (arg) {
-        is Number -> arg.toInt().toChar()
+        is Number -> arg.toChar()
         is Char   -> arg
         else      -> throw WrongTypeException(name, "Character or Number", arg)
     }
